pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-gud-samaritans"
version = "0.1.0"
description = "AI-powered open source contribution engine using nexus-agents orchestration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "William Zujkowski", email = "william@example.com"}
]
keywords = [
    "open-source",
    "contributions",
    "ai",
    "agents",
    "github",
    "automation",
    "mcp",
    "nexus-agents"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Version Control :: Git",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyGithub>=2.1.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "gitpython>=3.1.40",
    "aiosqlite>=0.19.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.12",
]
mcp = [
    "mcp>=0.1.0",
]

[project.scripts]
ggs = "git_gud_samaritans.cli:main"
git-gud-samaritans = "git_gud_samaritans.cli:main"

[project.urls]
Homepage = "https://github.com/williamzujkowski/git-gud-samaritans"
Documentation = "https://github.com/williamzujkowski/git-gud-samaritans#readme"
Repository = "https://github.com/williamzujkowski/git-gud-samaritans.git"
Issues = "https://github.com/williamzujkowski/git-gud-samaritans/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/git_gud_samaritans"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py310"

exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# E = pycodestyle errors, F = Pyflakes, I = isort, B = flake8-bugbear, UP = pyupgrade
select = ["E", "F", "I", "B", "UP", "C4", "SIM", "RUF"]
ignore = [
    "E501",  # Line too long - handled by formatter
]
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101"]  # Allow assert in tests

[tool.ruff.lint.isort]
known-first-party = ["git_gud_samaritans"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["github.*", "git.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/git_gud_samaritans"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
